requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
idna==2.5
oauthlib==2.0.2
pyperclip==1.5.27
python-dotenv==0.6.4
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
urllib3==1.21.1
